pyproject.toml
[project]
name = "proxy-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.2",
    "langgraph-cli[inmem]>=0.3.6",
    "langgraph-supervisor>=0.0.29",
    "langsmith[cli]>=0.4.8",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
]

[tool.black]
line-length = 89
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 89
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
